services/rpa-worker/requirements.txt
playwright
requests
python-dotenv
beautifulsoup4
loguru
